% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{combined_background}
\alias{combined_background}
\title{Background gene pairs.}
\usage{
data(combined_background)
}
\description{
This data includes the background Gene pairs used in DysPIA.
These 333,484 gene pairs from 8 public pathway databases above form a part of background sets, which is called the real background gene-pair set. In each gene pair,  the genes were arranged in ascending order based on gene symbols, that is, the first gene was smaller than the second gene.
700,000 gene pairs were randomly selected from 19297 genes in HGNC, and 349,915 of them in ascending order remained as the random background gene pair set. There were 982 gene pairs in common between the observed background and random background.
Finally, the combination of them with 682,417 gene pairs was set to be the background gene pair set, which is called combined background.
}
